fix: short url fixes
